Add tets for large number of items for hipcub::DeviceSelect::If

+++ hipcub/include/hipcub/backend/cub/device/device_select.hpp
@@ -53,7 +53,7 @@ public:
                               FlagIterator         d_flags,
                               OutputIteratorT      d_out,
                               NumSelectedIteratorT d_num_selected_out,
-                              int                  num_items,
+                              int64_t              num_items,
                               hipStream_t          stream = 0)
     {
         return hipCUDAErrorTohipError(::cub::DeviceSelect::Flagged(d_temp_storage,
@@ -77,7 +77,7 @@ public:
                               FlagIterator         d_flags,
                               OutputIteratorT      d_out,
                               NumSelectedIteratorT d_num_selected_out,
-                              int                  num_items,
+                              int64_t              num_items,
                               hipStream_t          stream,
                               bool                 debug_synchronous)
     {
@@ -99,7 +99,7 @@ public:
                               IteratorT            d_data,
                               FlagIterator         d_flags,
                               NumSelectedIteratorT d_num_selected_out,
-                              int                  num_items,
+                              int64_t              num_items,
                               hipStream_t          stream = 0)
     {
         return hipCUDAErrorTohipError(::cub::DeviceSelect::Flagged(d_temp_storage,
@@ -119,7 +119,7 @@ public:
                               IteratorT            d_data,
                               FlagIterator         d_flags,
                               NumSelectedIteratorT d_num_selected_out,
-                              int                  num_items,
+                              int64_t              num_items,
                               hipStream_t          stream,
                               bool                 debug_synchronous)
     {
@@ -143,7 +143,7 @@ public:
                          InputIteratorT       d_in,
                          OutputIteratorT      d_out,
                          NumSelectedIteratorT d_num_selected_out,
-                         int                  num_items,
+                         int64_t              num_items,
                          SelectOp             select_op,
                          hipStream_t          stream = 0)
     {
@@ -167,7 +167,7 @@ public:
                          InputIteratorT       d_in,
                          OutputIteratorT      d_out,
                          NumSelectedIteratorT d_num_selected_out,
-                         int                  num_items,
+                         int64_t              num_items,
                          SelectOp             select_op,
                          hipStream_t          stream,
                          bool                 debug_synchronous)
@@ -189,7 +189,7 @@ public:
                          size_t&              temp_storage_bytes,
                          IteratorT            d_data,
                          NumSelectedIteratorT d_num_selected_out,
-                         int                  num_items,
+                         int64_t              num_items,
                          SelectOp             select_op,
                          hipStream_t          stream = 0)
     {
@@ -208,7 +208,7 @@ public:
                          size_t&              temp_storage_bytes,
                          IteratorT            d_data,
                          NumSelectedIteratorT d_num_selected_out,
-                         int                  num_items,
+                         int64_t              num_items,
                          SelectOp             select_op,
                          hipStream_t          stream,
                          bool                 debug_synchronous)
@@ -230,7 +230,7 @@ public:
                              InputIteratorT       d_in,
                              OutputIteratorT      d_out,
                              NumSelectedIteratorT d_num_selected_out,
-                             int                  num_items,
+                             int64_t              num_items,
                              hipStream_t          stream = 0)
     {
         return hipCUDAErrorTohipError(::cub::DeviceSelect::Unique(d_temp_storage,
@@ -249,7 +249,7 @@ public:
                              InputIteratorT       d_in,
                              OutputIteratorT      d_out,
                              NumSelectedIteratorT d_num_selected_out,
-                             int                  num_items,
+                             int64_t              num_items,
                              hipStream_t          stream,
                              bool                 debug_synchronous)
     {
